% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t1}
\alias{butlin_t1}
\title{Table A1 of Butlin, Dixon and Lloyd 2014}
\format{
A tibble with six variables:
\describe{
  \item{year}{Financial year ending, eg. '2007' means the '2006-07' financial year. Sectoral shares prior to 1900 are calendar years.}
  \item{rgdp}{Real GDP, \$ million, constant 2010-11 prices}
  \item{ag_share}{Agriculture's share of GDP/GNP, current prices, in \%. Until 1861, includes pastoral and non-pastoral primary production.}
  \item{mining_share}{Mining's share of GDP/GNP, current prices, in \%}
  \item{manuf_share}{Manufacturing's share of GDP/GNP, current prices, in \%. Until 1861, includes private construction.}
  \item{other_share}{Other industries' share of GDP/GNP, current prices, in \%}
}
}
\source{
\url{https://doi.org/10.1017/CHO9781107445222.033}. See PDF for detailed notes and sources.
}
\usage{
butlin_t1
}
\description{
GDP and sectoral shares, 1801-2010.
}
\keyword{datasets}
