% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t2}
\alias{butlin_t2}
\title{Table A2 of Butlin, Dixon and Lloyd 2014}
\format{
A tibble with 10 variables:
\describe{
  \item{year}{Mix of calendar and financial year. Population is December each year; unemployment rate is financial year ended.}
  \item{males}{Male population}
  \item{females}{Female population}
  \item{persons}{Total population}
  \item{sex_ratio}{males \ females * 100}
  \item{net_migration_share_of_increase}{Share of population increase accounted for by net migration, per cent}
  \item{perc_pop_born_overseas}{Proportion of population born overseas, per cent}
  \item{labour_force}{Number of persons in the labour force, thousands}
  \item{unemp_rate}{Unemployment rate, as a percentage of the labour force}
  \item{part_rate}{Participation rate; labour force as a percentage of total (not working age) population}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t2
}
\description{
Population, migration, labour force and unemployment, 1800-2010.
}
\keyword{datasets}
