% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t3}
\alias{butlin_t3}
\title{Table A3 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010}
\format{
A tibble with 8 variables:
\describe{
  \item{year}{Prior to 1913: calendar year. 1914 onwards: financial year ended.}
  \item{exports}{Exports, in millions of dollars, nominal}
  \item{imports}{Imports, in millions of dollars, nominal}
  \item{balance_of_trade}{Balance of trade, exports minus imports, in millions of dollars, nominal}
  \item{openness_ratio}{Openness ratio as percentage of GDP; the sum of exports and imports as a percentage of GDP}
  \item{current_account_m}{Current account balance in millions of dollars, nominal}
  \item{current_account_percgdp}{Current account balance as a percentage of GDP}
  \item{terms_of_trade}{Terms of trade; index, 1966-67 = 100.}

}

A tibble with x variables:
\describe{
  \item{}{}
}

A tibble with x variables:
\describe{
  \item{}{}
}

A tibble with x variables:
\describe{
  \item{}{}
}

A tibble with x variables:
\describe{
  \item{}{}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.

https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.

https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.

https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.

https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t3

butlin_t3

butlin_t3

butlin_t3

butlin_t3
}
\description{
Table A3 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010

Table A4 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010

Table A3 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010

Table A3 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010

Table A3 of Butlin, Dixon and Lloyd 2014
Exports, imports, openness and current account deficit, 1822-2010
}
\keyword{datasets}
