% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t4}
\alias{butlin_t4}
\title{Table A4 of Butlin, Dixon and Lloyd 2014
Industry assistance, 1904-2005}
\format{
A tibble with 4 variables:
\describe{
  \item{year}{}
  \item{ave_nom_assist_ag}{Average nominal assistance to agriculture, per cent. Includes all covered agricultural commodities.}
  \item{ave_nom_assist_manuf}{Average nominal assistance to manufacturing, per cent. Average duty paid on dutiable imports.}
  \item{rra}{Relative rate of assistance. RRA is relative rate of assistance to manufacturing and is arrived at by computing the ratio of one plus the average nominal assistance to manufacturing to one plus the average nominal assistance to agriculture, subtracting one from the result and multiplying that by 100 to give a percentage.}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t4
}
\description{
Table A4 of Butlin, Dixon and Lloyd 2014
Industry assistance, 1904-2005
}
\keyword{datasets}
