% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t5}
\alias{butlin_t5}
\title{Table A5 of Butlin, Dixon and Lloyd 2014
Private capital formation, 1861-2010.}
\format{
A tibble with 3 variables:
\describe{
  \item{year}{Prior to 1900: calendar year. 1900 onwards: financial year ending.}
  \item{dwelling_inv_percgdp}{Private dwelling investment as per cent of GDP}
  \item{nondwelling_inv_percgdp}{Private non-dwelling investment as per cent of GDP}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t5
}
\description{
Table A5 of Butlin, Dixon and Lloyd 2014
Private capital formation, 1861-2010.
}
\keyword{datasets}
