% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t6}
\alias{butlin_t6}
\title{Table A6 of Butlin, Dixon and Lloyd 2014
Financial returns and selected asset prices, 1875-2010}
\format{
A tibble with 7 variables:
\describe{
  \item{year}{Year. For house prices, refers to June quarter. For other series, specific date is unclear.}
  \item{stock_index}{Stock index (1990 = 100).}
  \item{median_house_price}{Median house price in AUD, thousands, based on series for six capital cities.}
  \item{bill_yield}{Daily yield on 90 day bank-accepted bills}
  \item{bond_yield}{Daily yield on 10 year government bonds}
  \item{equity_premium}{Equity premium in per cent}
  \item{real_bond_yield}{Real bond yield is the bond yield minus the inflation rate}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t6
}
\description{
Table A6 of Butlin, Dixon and Lloyd 2014
Financial returns and selected asset prices, 1875-2010
}
\keyword{datasets}
