% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butlin_t7}
\alias{butlin_t7}
\title{Table A7 of Butlin, Dixon and Lloyd 2014
Living standards, wages, prices and housing affordability, 1801-2010}
\format{
A tibble with x variables:
\describe{
  \item{year}{Financial year ended}
  \item{rgdp_per_cap}{Real GDP per capita. RGDP in financial year divided by total population as at December in the fin year.}
  \item{awe}{Average weekly earnings, all employeed, dollars.}
  \item{gdp_def}{GDP deflator, 2010 = 100}
  \item{cpi}{Consumer price index, 1990  = 100}
  \item{real_prod_wage}{Real product wage (AWE deflated by GDP deflator)}
  \item{house_afford}{Housing affordability, median house prices (see \code{?butlin_t6}) divided by average weekly earnings.}
}
}
\source{
https://doi.org/10.1017/CHO9781107445222.033. See PDF for detailed notes and sources.
}
\usage{
butlin_t7
}
\description{
Table A7 of Butlin, Dixon and Lloyd 2014
Living standards, wages, prices and housing affordability, 1801-2010
}
\keyword{datasets}
