% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_lfs}
\alias{lr_lfs}
\title{Long-run labour force data
From Butlin}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 616 rows and 10 columns.
}
\source{
Butlin
}
\usage{
lr_lfs
}
\description{
Long-run labour force data
From Butlin
}
\keyword{datasets}
